* changed pk + fk in db to int(10)

+++ config/db/tables.sql
@@ -3,7 +3,7 @@
 --
 
 CREATE TABLE `ext_contact_person` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL,
   `id_person_create` smallint(5) unsigned NOT NULL DEFAULT '0',
@@ -29,10 +29,10 @@ CREATE TABLE `ext_contact_person` (
 --
 
 CREATE TABLE `ext_contact_company` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(2) NOT NULL DEFAULT '0',
   `title` tinytext NOT NULL,
   `shortname` tinytext NOT NULL,
@@ -50,10 +50,10 @@ CREATE TABLE `ext_contact_company` (
 --
 
 CREATE TABLE `ext_contact_address` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(2) NOT NULL,
   `id_addresstype` tinyint(1) unsigned NOT NULL DEFAULT '0',
   `id_country` tinyint(1) unsigned NOT NULL DEFAULT '0',
@@ -76,10 +76,10 @@ CREATE TABLE `ext_contact_address` (
 --
 
 CREATE TABLE `ext_contact_contactinfo` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(1) NOT NULL DEFAULT '0',
   `id_contactinfotype` tinytext NOT NULL,
   `info` tinytext NOT NULL,
@@ -94,10 +94,10 @@ CREATE TABLE `ext_contact_contactinfo` (
 --
 
 CREATE TABLE `ext_contact_contactinfotype` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(11) NOT NULL,
   `date_update` int(11) NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(1) NOT NULL DEFAULT '0',
   `category` smallint(5) unsigned NOT NULL,
   `key` varchar(20) NOT NULL,
@@ -112,10 +112,10 @@ CREATE TABLE `ext_contact_contactinfotype` (
 --
 
 CREATE TABLE `ext_contact_jobtype` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(11) NOT NULL,
   `date_update` int(11) NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(1) NOT NULL DEFAULT '0',
   `title` varchar(48) NOT NULL,
   PRIMARY KEY (`id`)
@@ -128,7 +128,7 @@ CREATE TABLE `ext_contact_jobtype` (
 --
 
 CREATE TABLE `ext_contact_mm_company_address` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `id_company` int(10) unsigned NOT NULL DEFAULT '0',
   `id_address` int(10) unsigned NOT NULL,
   PRIMARY KEY (`id`)
@@ -141,7 +141,7 @@ CREATE TABLE `ext_contact_mm_company_address` (
 --
 
 CREATE TABLE `ext_contact_mm_company_contactinfo` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `id_company` int(10) unsigned NOT NULL DEFAULT '0',
   `id_contactinfo` int(10) unsigned NOT NULL,
   PRIMARY KEY (`id`)
@@ -154,11 +154,11 @@ CREATE TABLE `ext_contact_mm_company_contactinfo` (
 --
 
 CREATE TABLE `ext_contact_mm_company_person` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `id_company` int(10) unsigned NOT NULL DEFAULT '0',
-  `id_person` smallint(5) unsigned NOT NULL,
-  `id_workaddress` smallint(6) NOT NULL,
-  `id_jobtype` smallint(6) NOT NULL,
+  `id_person` int(10) unsigned NOT NULL,
+  `id_workaddress` int(10) NOT NULL,
+  `id_jobtype` int(10) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=MyISAM  DEFAULT CHARSET=utf8;
 
@@ -182,8 +182,8 @@ CREATE TABLE `ext_contact_mm_person_address` (
 --
 
 CREATE TABLE `ext_contact_mm_person_contactinfo` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `id_person` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id` int(10) NOT NULL AUTO_INCREMENT,
+  `id_person` int(10) unsigned NOT NULL DEFAULT '0',
   `id_contactinfo` int(10) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
@@ -195,8 +195,8 @@ CREATE TABLE `ext_contact_mm_person_contactinfo` (
 --
 
 CREATE TABLE `ext_contact_mm_person_role` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `id_person` smallint(5) unsigned NOT NULL DEFAULT '0',
-  `id_role` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id` int(10) NOT NULL AUTO_INCREMENT,
+  `id_person` int(10) unsigned NOT NULL DEFAULT '0',
+  `id_role` int(10) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
